movie/requirements.txt
# Movie to PubSub Events Publisher Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud dependencies
google-cloud-pubsub>=2.18.0
google-cloud-core>=2.3.0
google-auth>=2.17.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0